chore(service): move Color and ValidationException to util package #30
