+++ src/main/java/pl/gawor/tayckner/taycknerbackend/service/facade/CategoryFacade.java
@@ -3,6 +3,8 @@ package pl.gawor.tayckner.taycknerbackend.service.facade;
 import org.springframework.stereotype.Component;
 import pl.gawor.tayckner.taycknerbackend.core.model.CategoryModel;
 import pl.gawor.tayckner.taycknerbackend.core.model.UserModel;
+import pl.gawor.tayckner.taycknerbackend.service.facade.util.Color;
+import pl.gawor.tayckner.taycknerbackend.service.facade.util.ValidationException;
 import pl.gawor.tayckner.taycknerbackend.service.service.CategoryService;
 import pl.gawor.tayckner.taycknerbackend.service.service.UserService;
 import pl.gawor.tayckner.taycknerbackend.web.response.Response;
