+++ src/main/java/pl/gawor/tayckner/taycknerbackend/service/facade/UserFacade.java
@@ -2,6 +2,7 @@ package pl.gawor.tayckner.taycknerbackend.service.facade;
 
 import org.springframework.stereotype.Component;
 import pl.gawor.tayckner.taycknerbackend.core.model.UserModel;
+import pl.gawor.tayckner.taycknerbackend.service.facade.util.ValidationException;
 import pl.gawor.tayckner.taycknerbackend.service.service.UserService;
 import pl.gawor.tayckner.taycknerbackend.web.response.Response;
 import pl.gawor.tayckner.taycknerbackend.web.response.ResponseStatus;
