+++ src/main/java/pl/gawor/tayckner/taycknerbackend/service/facade/util/ValidationException.java
@@ -1,4 +1,4 @@
-package pl.gawor.tayckner.taycknerbackend.service.facade;
+package pl.gawor.tayckner.taycknerbackend.service.facade.util;
 
 public class ValidationException extends Exception {
     public ValidationException() {
